refactor: move feature domain to domain module
