+++ domain/src/main/java/com/emotionstorage/domain/useCase/user/ValidateNicknameUseCase.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.tutorial.domain
+package com.emotionstorage.domain.useCase.user
 
 import com.emotionstorage.domain.common.DataState
 import javax.inject.Inject
